backend/requirements.txt
django>=5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
channels==4.1.0
channels-redis==4.2.0